backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
anthropic==0.28.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
pydantic==2.7.1
yfinance==0.2.40
beautifulsoup4==4.12.3
lxml==5.2.2
python-dateutil==2.9.0
aiohttp==3.9.5
